requirements.txt
Pillow>=10.0.0
ttkbootstrap>=1.4.7
